move QtSettingsStorage to separate file
