+++ libosmscout-client-qt/src/osmscoutclientqt/OSMScoutQt.cpp
@@ -23,6 +23,7 @@
 
 #include <osmscoutclientqt/OSMScoutQt.h>
 #include <osmscoutclientqt/Settings.h>
+#include <osmscoutclientqt/QtSettingsStorage.h>
 #include <osmscoutclientqt/DBThread.h>
 #include <osmscoutclientqt/MapWidget.h>
 #include <osmscoutclientqt/ElevationChartWidget.h>
