+++ libosmscout-client-qt/src/osmscoutclientqt/Settings.cpp
@@ -34,88 +34,6 @@
 
 namespace osmscout {
 
-QtSettingsStorage::QtSettingsStorage(QSettings *providedStorage):
-  storage(providedStorage)
-{
-  if (storage==nullptr){
-    storage=new QSettings(this);
-  } else {
-    providedStorage->setParent(this);
-  }
-}
-
-void QtSettingsStorage::SetValue(const std::string &key, double d)
-{
-  storage->setValue(QString::fromStdString(key), d);
-}
-
-void QtSettingsStorage::SetValue(const std::string &key, uint32_t i)
-{
-  storage->setValue(QString::fromStdString(key), i);
-}
-
-void QtSettingsStorage::SetValue(const std::string &key, const std::string &str)
-{
-  storage->setValue(QString::fromStdString(key), QString::fromStdString(str));
-}
-
-void QtSettingsStorage::SetValue(const std::string &key, bool b)
-{
-  storage->setValue(QString::fromStdString(key), b);
-}
-
-void QtSettingsStorage::SetValue(const std::string &key, std::vector<char> bytes)
-{
-  storage->setValue(QString::fromStdString(key), QByteArray(bytes.data(), bytes.size()));
-}
-
-double QtSettingsStorage::GetDouble(const std::string &key, double defaultValue)
-{
-  bool ok;
-  double d = storage->value(QString::fromStdString(key), defaultValue).toDouble(&ok);
-  if (!ok) {
-    return defaultValue;
-  }
-  return d;
-}
-
-uint32_t QtSettingsStorage::GetUInt(const std::string &key, uint32_t defaultValue)
-{
-  bool ok;
-  uint32_t i = storage->value(QString::fromStdString(key), defaultValue).toUInt(&ok);
-  if (!ok) {
-    return defaultValue;
-  }
-  return i;
-}
-
-std::string QtSettingsStorage::GetString(const std::string &key, const std::string &defaultValue)
-{
-  return storage->value(QString::fromStdString(key), QString::fromStdString(defaultValue)).toString().toStdString();
-}
-
-bool QtSettingsStorage::GetBool(const std::string &key, bool defaultValue)
-{
-  return storage->value(QString::fromStdString(key), defaultValue).toBool();
-}
-
-std::vector<char> QtSettingsStorage::GetBytes(const std::string &key)
-{
-  QByteArray arr = storage->value(QString::fromStdString(key)).toByteArray();
-  return std::vector<char>(arr.data(), arr.data() + arr.size());
-}
-
-std::vector<std::string> QtSettingsStorage::Keys(const std::string &prefix)
-{
-  std::vector<std::string> result;
-  storage->beginGroup(QString::fromStdString(prefix));
-  for (const QString& key:storage->allKeys()){
-    result.push_back(prefix + key.toStdString());
-  }
-  storage->endGroup();
-  return result;
-}
-
 Settings::Settings(SettingsStoragePtr storage):
   storage(storage)
 {
